split library up further to help with low memory boards
