+++ adafruit_led_animation/animation/colorcycle.py
@@ -1,0 +1,73 @@
+# The MIT License (MIT)
+#
+# Copyright (c) 2019-2020 Roy Hooper
+# Copyright (c) 2020 Kattni Rembor for Adafruit Industries
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+"""
+`adafruit_led_animation.animation.colorcycle`
+================================================================================
+
+TODO
+
+* Author(s): Roy Hooper, Kattni Rembor
+
+"""
+
+from adafruit_led_animation.animation import Animation
+from adafruit_led_animation.color import RAINBOW
+
+
+class ColorCycle(Animation):
+    """
+    Animate a sequence of one or more colors, cycling at the specified speed.
+
+    :param pixel_object: The initialised LED object.
+    :param float speed: Animation speed in seconds, e.g. ``0.1``.
+    :param colors: A list of colors to cycle through in ``(r, g, b)`` tuple, or ``0x000000`` hex
+                   format. Defaults to a rainbow color cycle.
+    """
+
+    def __init__(self, pixel_object, speed, colors=RAINBOW, name=None):
+        self.colors = colors
+        super().__init__(pixel_object, speed, colors[0], name=name)
+        self._generator = self._color_generator()
+        next(self._generator)
+
+    cycle_complete_supported = True
+
+    def draw(self):
+        self.pixel_object.fill(self.color)
+        self.show()
+        next(self._generator)
+
+    def _color_generator(self):
+        index = 0
+        while True:
+            self._color = self.colors[index]
+            yield
+            index = (index + 1) % len(self.colors)
+            if index == 0:
+                self.cycle_complete()
+
+    def reset(self):
+        """
+        Resets to the first color.
+        """
+        self._generator = self._color_generator()
