+++ adafruit_led_animation/animation/rainbowchase.py
@@ -1,0 +1,71 @@
+# The MIT License (MIT)
+#
+# Copyright (c) 2019-2020 Roy Hooper
+# Copyright (c) 2020 Kattni Rembor for Adafruit Industries
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+# THE SOFTWARE.
+"""
+`adafruit_led_animation.animation.rainbowchase`
+================================================================================
+
+TODO
+
+* Author(s): Roy Hooper, Kattni Rembor
+
+"""
+
+from adafruit_led_animation.color import colorwheel
+from adafruit_led_animation.animation.chase import Chase
+
+
+class RainbowChase(Chase):
+    """
+    Chase pixels in one direction, like a theater marquee but with rainbows!
+
+    :param pixel_object: The initialised LED object.
+    :param float speed: Animation speed rate in seconds, e.g. ``0.1``.
+    :param color: Animation color in ``(r, g, b)`` tuple, or ``0x000000`` hex format.
+    :param size: Number of pixels to turn on in a row.
+    :param spacing: Number of pixels to turn off in a row.
+    :param reverse: Reverse direction of movement.
+    :param wheel_step: How many colors to skip in `colorwheel` per bar (default 8)
+    """
+
+    # pylint: disable=too-many-arguments
+    def __init__(
+        self,
+        pixel_object,
+        speed,
+        size=2,
+        spacing=3,
+        reverse=False,
+        name=None,
+        wheel_step=8,
+    ):
+        self._num_colors = 256 // wheel_step
+        self._colors = [colorwheel(n % 256) for n in range(0, 512, wheel_step)]
+        self._color_idx = 0
+        super().__init__(pixel_object, speed, 0, size, spacing, reverse, name)
+
+    def bar_color(self, n, pixel_no=0):
+        return self._colors[self._color_idx - n]
+
+    def cycle_complete(self):
+        self._color_idx = (self._color_idx + self._direction) % len(self._colors)
+        super().cycle_complete()
