+++ adafruit_led_animation/color.py
@@ -45,8 +45,8 @@ RAINBOW = (RED, ORANGE, YELLOW, GREEN, BLUE, PURPLE)
 
 
 try:
-    # Backwards compat for 5.3.0 and prior
     try:
+        # Backwards compat for 5.3.0 and prior
         from _pixelbuf import colorwheel  # pylint: disable=unused-import
     except ImportError:
         from _pixelbuf import wheel as colorwheel  # pylint: disable=unused-import
