+++ examples/led_animation_all_animations.py
@@ -9,8 +9,8 @@ This example does not work on SAMD21 (M0) boards.
 import board
 import neopixel
 
-import adafruit_led_animation.rainbow
-import adafruit_led_animation.sparkle
+import adafruit_led_animation.animation.blink
+import adafruit_led_animation.animation.sparklepulse
 from adafruit_led_animation import animation
 from adafruit_led_animation.sequence import AnimationSequence
 from adafruit_led_animation.color import PURPLE, WHITE, AMBER, JADE
@@ -22,22 +22,22 @@ pixel_num = 32
 
 pixels = neopixel.NeoPixel(pixel_pin, pixel_num, brightness=0.2, auto_write=False)
 
-blink = animation.Blink(pixels, speed=0.5, color=JADE)
+blink = adafruit_led_animation.animation.animation.blink.Blink(pixels, speed=0.5, color=JADE)
 comet = animation.Comet(pixels, speed=0.01, color=PURPLE, tail_length=10, bounce=True)
 chase = animation.Chase(pixels, speed=0.1, size=3, spacing=6, color=WHITE)
 pulse = animation.Pulse(pixels, speed=0.1, period=3, color=AMBER)
-sparkle = adafruit_led_animation.sparkle.Sparkle(
+sparkle = adafruit_led_animation.animation.animation.sparkle.Sparkle(
     pixels, speed=0.1, color=PURPLE, num_sparkles=10
 )
-solid = animation.Solid(pixels, color=JADE)
-rainbow = adafruit_led_animation.rainbow.Rainbow(pixels, speed=0.1, period=2)
-sparkle_pulse = adafruit_led_animation.sparkle.SparklePulse(
+solid = adafruit_led_animation.animation.animation.solid.Solid(pixels, color=JADE)
+rainbow = adafruit_led_animation.animation.animation.rainbow.Rainbow(pixels, speed=0.1, period=2)
+sparkle_pulse = adafruit_led_animation.animation.animation.sparklepulse.SparklePulse(
     pixels, speed=0.1, period=3, color=JADE
 )
-rainbow_comet = adafruit_led_animation.rainbow.RainbowComet(
+rainbow_comet = adafruit_led_animation.animation.animation.rainbowcomet.RainbowComet(
     pixels, speed=0.1, tail_length=7, bounce=True
 )
-rainbow_chase = adafruit_led_animation.rainbow.RainbowChase(
+rainbow_chase = adafruit_led_animation.animation.animation.rainbowchase.RainbowChase(
     pixels, speed=0.1, size=3, spacing=2, wheel_step=8
 )
